xRevert "Stuff"
This reverts commit 9b9f4f3fc81410cbefcd82308178155d0e7a11cf.

+++ src/main/java/cn/nukkit/network/protocol/RemoveEntityPacket.java
@@ -18,7 +18,7 @@ public class RemoveEntityPacket extends DataPacket {
 
     @Override
     public void decode() {
-        this.entityNetworkId = (int) this.getUnsignedVarInt();
+    	this.entityNetworkId = this.getUnsignedVarInt();
     }
 
     @Override
